pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stegano-anomaly"
version = "0.1.0"
description = "Steganographic analysis via engineered features + anomaly detection baselines"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Your Name"}]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pillow>=10.0",
  "scipy>=1.10",
  "scikit-learn>=1.3",
  "joblib>=1.3",
  "typer>=0.9",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.3.0",
  "tqdm>=4.66",
  "pyinstaller>=6.0",
]

[project.scripts]
stegano-anomaly = "stegano_anomaly.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
